requirements.txt
# --- Core ---
numpy>=1.24.4,<1.27.0
torch==1.12.1

# --- Dependencies ---
h5py>=3.10.0
ipython>=8.12.0
ipywidgets>=8.0.0
matplotlib>=3.8.0
MDAnalysis>=2.7.0
PyYAML>=6.0.0
ray[default]>=2.40.0
scipy>=1.10.0  
tfrecord>=1.14.0
tqdm>=4.66.0
webdataset>=0.2.100
jupyterlab>=4.0.0
